feat: 폴더 구조 정리, WeekNavigator 스타일 파일 분리
